fix: split test in attempt to evade CI badness
