+++ packages/SwingSet/test/vat-admin/terminate/test-terminate-non-critical.js
@@ -67,37 +67,6 @@ async function doTerminateNonCritical(
   t.is(postProbe, undefined);
 }
 
-async function doTerminateCritical(t, deadVatID, mode, dynamic) {
-  const configPath = new URL('swingset-terminate.json', import.meta.url)
-    .pathname;
-  const config = await loadSwingsetConfigFile(configPath);
-  const hostStorage = provideHostStorage();
-  const controller = await buildVatController(config, [mode, true, dynamic], {
-    ...t.context.data,
-    hostStorage,
-  });
-  t.is(controller.kpStatus(controller.bootstrapResult), 'unresolved');
-  const preProbe = hostStorage.kvStore.get(`${deadVatID}.options`);
-  if (!dynamic) {
-    t.truthy(preProbe);
-  }
-
-  const err = await t.throwsAsync(() => controller.run());
-  const body = JSON.parse(err.message);
-  if (typeof body === 'string') {
-    t.is(body, mode);
-  } else {
-    t.is(body['@qclass'], 'error');
-    t.is(body.message, mode);
-  }
-  t.is(
-    controller.kpStatus(controller.bootstrapResult),
-    dynamic ? 'unresolved' : 'fulfilled',
-  );
-  const postProbe = hostStorage.kvStore.get(`${deadVatID}.options`);
-  t.is(postProbe, undefined);
-}
-
 test('terminate (dynamic, non-critical)', async t => {
   await doTerminateNonCritical(
     t,
@@ -109,12 +78,6 @@ test('terminate (dynamic, non-critical)', async t => {
 });
 
 // no test 'terminate (static, non-critical)' because static vats can't be killed
-
-test('terminate (dynamic, critical)', async t => {
-  await doTerminateCritical(t, 'v8', 'kill', true);
-});
-
-// no test 'terminate (static, critical)' because static vats can't be killed
 
 test('exit happy path simple result (dynamic, non-critical)', async t => {
   await doTerminateNonCritical(
@@ -134,14 +97,6 @@ test('exit happy path simple result (static, non-critical)', async t => {
     false,
     'done result happy (Error=false)',
   );
-});
-
-test('exit happy path simple result (dynamic, critical)', async t => {
-  await doTerminateCritical(t, 'v8', 'happy', true);
-});
-
-test('exit happy path simple result (static, critical)', async t => {
-  await doTerminateCritical(t, 'v3', 'happy', false);
 });
 
 test('exit happy path complex result (dynamic, non-critical)', async t => {
@@ -164,14 +119,6 @@ test('exit happy path complex result (static, non-critical)', async t => {
   );
 });
 
-test('exit happy path complex result (dynamic, critical)', async t => {
-  await doTerminateCritical(t, 'v8', 'exceptionallyHappy', true);
-});
-
-test('exit happy path complex result (static, critical)', async t => {
-  await doTerminateCritical(t, 'v3', 'exceptionallyHappy', false);
-});
-
 test('exit sad path simple result (dynamic, non-critical)', async t => {
   await doTerminateNonCritical(
     t,
@@ -190,14 +137,6 @@ test('exit sad path simple result (static, non-critical)', async t => {
     false,
     'done exception sad (Error=false)',
   );
-});
-
-test('exit sad path simple result (dynamic, critical)', async t => {
-  await doTerminateCritical(t, 'v8', 'sad', true);
-});
-
-test('exit sad path simple result (static, critical)', async t => {
-  await doTerminateCritical(t, 'v3', 'sad', false);
 });
 
 test('exit sad path complex result (dynamic, non-critical)', async t => {
@@ -220,14 +159,6 @@ test('exit sad path complex result (static, non-critical)', async t => {
   );
 });
 
-test('exit sad path complex result (dynamic, critical)', async t => {
-  await doTerminateCritical(t, 'v8', 'exceptionallySad', true);
-});
-
-test('exit sad path complex result (static, critical)', async t => {
-  await doTerminateCritical(t, 'v3', 'exceptionallySad', false);
-});
-
 test('exit happy path with ante-mortem message (dynamic, non-critical)', async t => {
   await doTerminateNonCritical(
     t,
@@ -248,14 +179,6 @@ test('exit happy path with ante-mortem message (static, non-critical)', async t 
     'done result happyTalkFirst (Error=false)',
     ['GOT QUERY not dead quite yet'],
   );
-});
-
-test('exit happy path with ante-mortem message (dynamic, critical)', async t => {
-  await doTerminateCritical(t, 'v8', 'happyTalkFirst', true);
-});
-
-test('exit happy path with ante-mortem message (static, critical)', async t => {
-  await doTerminateCritical(t, 'v3', 'happyTalkFirst', false);
 });
 
 test('exit sad path with ante-mortem message (dynamic, non-critical)', async t => {
@@ -282,14 +205,6 @@ test('exit sad path with ante-mortem message (static, non-critical)', async t =>
     // *not* be observed here
     // ['GOT QUERY not dead quite yet'],
   );
-});
-
-test('exit sad path with ante-mortem message (dynamic, critical)', async t => {
-  await doTerminateCritical(t, 'v8', 'sadTalkFirst', true);
-});
-
-test('exit sad path with ante-mortem message (static, critical)', async t => {
-  await doTerminateCritical(t, 'v3', 'sadTalkFirst', false);
 });
 
 test('exit with presence', async t => {
@@ -347,16 +262,6 @@ test.serial('dispatches to the dead do not harm kernel', async t => {
       'm: live 2 failed: Error: vat terminated',
     ]);
   }
-});
-
-test('invalid criticalVatKey causes vat creation to fail', async t => {
-  const configPath = new URL('swingset-bad-vat-key.json', import.meta.url)
-    .pathname;
-  const config = await loadSwingsetConfigFile(configPath);
-  const controller = await buildVatController(config, [], t.context.data);
-  await t.throwsAsync(() => controller.run(), {
-    message: /invalid criticalVatKey/,
-  });
 });
 
 test('dead vat state removed', async t => {
